pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fpv-vision"
version = "0.1.0"
description = "FPV drone component detection using YOLO"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "ultralytics>=8.0",
    "torch>=2.0",
    "torchvision>=0.15",
    "opencv-python>=4.8",
    "typer[all]>=0.12",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "PyYAML>=6.0",
    "supervision>=0.18",
    "numpy>=1.24",
    "matplotlib>=3.7",
]

[project.scripts]
fpv-vision = "fpv_vision.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
